+++ storybook/lib/module/widget_showcase/presentation/screen/input/input/_widget/_label_behavior_option.dart
@@ -17,14 +17,12 @@ final class _LabelBehaviorOption extends StatelessWidget {
     return MyoroSingleSelectionDropdown<FloatingLabelBehavior>(
       configuration: MyoroSingleSelectionDropdownConfiguration(
         label: 'Label behavior',
-        selectedItemBuilder:
-            (behavior) =>
-                _behaviors.entries
-                    .firstWhere(
-                      (entry) => entry.value == behavior,
-                      orElse: () => const MapEntry('custom', FloatingLabelBehavior.auto),
-                    )
-                    .key,
+        selectedItemBuilder: (behavior) => _behaviors.entries
+            .firstWhere(
+              (entry) => entry.value == behavior,
+              orElse: () => const MapEntry('custom', FloatingLabelBehavior.auto),
+            )
+            .key,
         onChanged: (behavior) => viewModel.state.labelBehavior = behavior,
         menuConfiguration: MyoroSingleMenuConfiguration(
           request: () => _behaviors.values.toSet(),