+++ storybook/lib/module/widget_showcase/presentation/screen/slider/slider/view_model/myoro_sliders_widget_showcase_screen_view_model.dart
@@ -22,8 +22,9 @@ final class MyoroSlidersWidgetShowcaseScreenViewModel {
       min: state.min,
       max: state.max,
       value: state.sliderValue,
-      currentValueText:
-          state.currentValueTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.currentValueTextDefaultValue,
+      currentValueText: state.currentValueTextEnabled
+          ? formattedValue
+          : MyoroSliderBaseConfiguration.currentValueTextDefaultValue,
       maxValueText: state.maxValueTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.maxValueTextDefaultValue,
       footerText: state.footerTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.footerTextDefaultValue,
       onChanged: (value) => state.sliderValue = value,
@@ -40,8 +41,9 @@ final class MyoroSlidersWidgetShowcaseScreenViewModel {
       min: state.min,
       max: state.max,
       values: state.rangeSliderValues,
-      currentValueText:
-          state.currentValueTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.currentValueTextDefaultValue,
+      currentValueText: state.currentValueTextEnabled
+          ? formattedValue
+          : MyoroSliderBaseConfiguration.currentValueTextDefaultValue,
       maxValueText: state.maxValueTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.maxValueTextDefaultValue,
       footerText: state.footerTextEnabled ? formattedValue : MyoroSliderBaseConfiguration.footerTextDefaultValue,
       onChanged: (value) => state.rangeSliderValues = value,