+++ storybook/lib/module/widget_showcase/presentation/screen/table/view_model/myoro_table_widget_showcase_screen_view_model.dart
@@ -8,19 +8,18 @@ final class MyoroTableWidgetShowcaseScreenViewModel {
   MyoroTableConfiguration<String> buildConfiguration(BuildContext context) {
     final request = List.generate(faker.randomGenerator.integer(50), (int index) => 'Item #$index').toSet;
 
-    final columns =
-        MyoroTableColumnWidthConfigurationEnum.values.map((value) {
-          final name = value.name.capitalized;
+    final columns = MyoroTableColumnWidthConfigurationEnum.values.map((value) {
+      final name = value.name.capitalized;
 
-          return MyoroTableColumn(
-            tooltipMessage: 'Tooltip of the $name column.',
-            widthConfiguration: MyoroTableColumnWidthConfiguration(
-              typeEnum: value,
-              fixedWidth: value.isFixed ? faker.randomGenerator.decimal(scale: 200, min: 50) : null,
-            ),
-            child: Text(name),
-          );
-        }).toList();
+      return MyoroTableColumn(
+        tooltipMessage: 'Tooltip of the $name column.',
+        widthConfiguration: MyoroTableColumnWidthConfiguration(
+          typeEnum: value,
+          fixedWidth: value.isFixed ? faker.randomGenerator.decimal(scale: 200, min: 50) : null,
+        ),
+        child: Text(name),
+      );
+    }).toList();
 
     return MyoroTableConfiguration(
       request: request,
@@ -41,15 +40,16 @@ final class MyoroTableWidgetShowcaseScreenViewModel {
     return MyoroTableRow(
       onTapDown: (_) => onTapDown(context, item),
       onTapUp: (_) => onTapUp(context, item),
-      cells:
-          MyoroTableColumnWidthConfigurationEnum.values.map((value) {
-            return Text('$item\'s value.name.capitalized cell.');
-          }).toList(),
+      cells: MyoroTableColumnWidthConfigurationEnum.values.map((value) {
+        return Text('$item\'s value.name.capitalized cell.');
+      }).toList(),
     );
   }
 
   /// Generic function to display a [MyoroSnackBar].
   void _showSnackBar(BuildContext context, String message) {
-    context.showSnackBar(snackBar: MyoroSnackBar(configuration: MyoroSnackBarConfiguration(message: message)));
+    context.showSnackBar(
+      snackBar: MyoroSnackBar(configuration: MyoroSnackBarConfiguration(message: message)),
+    );
   }
 }