+++ storybook/lib/module/widget_showcase/presentation/widget/widget_showcase_option/padding/_widget/_selection_input_pair.dart
@@ -36,8 +36,12 @@ final class _SelectionInputPair extends StatelessWidget {
     return Row(
       spacing: widgetShowcaseThemeExtension.spacing,
       children: [
-        Expanded(child: _SelectionInput(label: leftLabel, onChanged: leftOnChanged, value: leftValue)),
-        Expanded(child: _SelectionInput(label: rightLabel, onChanged: rightOnChanged, value: rightValue)),
+        Expanded(
+          child: _SelectionInput(label: leftLabel, onChanged: leftOnChanged, value: leftValue),
+        ),
+        Expanded(
+          child: _SelectionInput(label: rightLabel, onChanged: rightOnChanged, value: rightValue),
+        ),
       ],
     );
   }